Annotation Refactor - move actions to Annotations feature dir
