+++ src/annotations/README.md
@@ -1,0 +1,5 @@
+# Annotations
+
+An Annotation is a saved highlight, that may or may not contain a note against it.
+
+In this feature folder (`src/annotations`) exists...
