+++ src/annotations/index.ts
@@ -3,6 +3,8 @@ import { renderHighlights } from 'src/highlighting/ui/highlight-interactions'
 import { createHighlight } from 'src/highlighting/ui'
 import { toggleSidebarOverlay } from 'src/sidebar-overlay/utils'
 
+// TODO (ch - annotations): If this function is doing to work of getting all the annotations for a tab,
+// TODO (ch - annotations): it can at least also provide it for the annotations in the side bar.
 export const fetchAnnotationsAndHighlight = async () => {
     const annotationList = await remoteFunction('getAllAnnotationsByUrl')({
         url: window.location.href,
